KS-1-5 time in space ( refactor )

+++ src/main/kotlin/lesson_1/lesson1_task4.kt
@@ -2,14 +2,12 @@ package org.example.lesson_1
 
 fun main() {
 
-    // Решение задачи №4
-
     // Данные полета Юрия Гагарина
 
     val length: Long = 40_868_600_000
-    val age: Short = 27
+    val age: Byte = 27
     val partOfDay: Float = 0.075f
-    val seconds: Int = 6480
+    val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
     println("$length \n$age \n$partOfDay \n$seconds \n$partOfYear \n$apogee")
